cleanup of the SDKBeaconAction Object
Conflicts:
	SensorbergSDK.xcodeproj/project.pbxproj
	Tests/Specs/SBSDKBeaconEventResponseObjectSpec.m
