+++ SensorbergSDK/SBSDKBeaconEventResponseObject.m
@@ -62,7 +62,7 @@
 
             for (id eachObject in (NSArray *)_response) {
                 if ([eachObject isKindOfClass:[NSDictionary class]]) {
-                    [newEvents addObject:[[SBSDKBeaconAction alloc] initWithAction:(NSDictionary *)eachObject]];
+                    [newEvents addObject:[[SBSDKBeaconAction alloc] initWithJSONDictionary:(NSDictionary *) eachObject]];
                 }
             }
